+++ inlang/source-code/paraglide/paraglide-js/src/cli/index.ts
@@ -1,10 +1,9 @@
 import { Command } from "commander"
-import { compileCommand2 } from "./commands/compile2/command.js"
-// import { initCommand2 } from "./commands/init2/command.js"
+import { compileCommand } from "./commands/compile/command.js"
 
 export const cli = new Command()
 	.name("paraglide-js")
-	.addCommand(compileCommand2)
+	.addCommand(compileCommand)
 	// .addCommand(initCommand2)
 	.showHelpAfterError()
 	.version(PARJS_PACKAGE_VERSION)
